requirements.txt
Django==1.10.1
dj-database-url==0.4.1
dj-static==0.0.6
django-appconf==1.0.2
django-compressor==2.1
djangorestframework==3.4.6
drf-nested-routers==0.11.1
gunicorn==19.6.0
six==1.10.0
static3==0.7.0
wsgiref==0.1.2
django-secure==1.0.1
MySQL-python==1.2.5
django-twilio==0.8.0
twilio==5.7.0
django-csp==3.1
django-registration==2.2
django-sslserver==0.19
requests==2.13.0